+++ backend/router/webtoon_router.py
@@ -1,12 +1,9 @@
 from fastapi import APIRouter, Depends, HTTPException, status
-from models import crud, models, schemas
-from models.crud import pwd_context
-from datetime import timedelta, datetime
+from models import schemas
 from sqlalchemy.orm import Session
 from models.database import get_db
-from typing import List
-from jose import jwt
 from auth.oauth import get_current_user
+from util import webtoon_util
 
 
 api_webtoon = APIRouter(prefix="/api/webtoon")
@@ -21,14 +18,14 @@ api_webtoon = APIRouter(prefix="/api/webtoon")
 
 @api_webtoon.post('/create')
 def create_webtoon(request: schemas.WebtoonCreate, db: Session = Depends(get_db), current_user: schemas.User = Depends(get_current_user)):
-    db_webtoon = crud.get_webtoon_by_webtoon_name_and_userId(db, webtoon_name=request.webtoonName, user_id=current_user['userId'])
+    db_webtoon = webtoon_util.get_webtoon_by_webtoon_name_and_userId(db, webtoon_name=request.webtoonName, user_id=current_user['userId'])
     if db_webtoon:
         raise HTTPException(status_code=400, detail="Bad Request: Webtoon already registered")
-    return crud.create_webtoon(db, webtoon=request, user=current_user)
+    return webtoon_util.create_webtoon(db, webtoon=request, user=current_user)
 
 @api_webtoon.get('/list')
 def read_webtoon_list(db: Session = Depends(get_db), current_user: schemas.User = Depends(get_current_user)):
-    webtoons = crud.get_webtoon_list_by_user_id(db, current_user['userId'])
+    webtoons = webtoon_util.get_webtoon_list_by_user_id(db, current_user['userId'])
     formatted_webtoons = []
     for webtoon in webtoons:
         formatted_webtoons.append({
@@ -40,7 +37,7 @@ def read_webtoon_list(db: Session = Depends(get_db), current_user: schemas.User 
 
 @api_webtoon.delete('/delete/{webtoon_name}')
 def delete_webtoon(request: schemas.WebtoonCreate, db: Session = Depends(get_db), current_user: schemas.User = Depends(get_current_user)):
-    db_webtoon = crud.get_webtoon_by_webtoon_name_and_userId(db, webtoon_name=request.webtoonName, user_id=current_user['userId'])
+    db_webtoon = webtoon_util.get_webtoon_by_webtoon_name_and_userId(db, webtoon_name=request.webtoonName, user_id=current_user['userId'])
     if db_webtoon:
         db.delete(db_webtoon)
         db.commit()
